Quick and dirty basics
Implement an action that allows the moderator to set the removal message body
Implement an action that allows the moderator to set the ban message body
Implement an action that allows the moderator to configure a custom macro
Implement an action to remove a post w/ macro response
Implement an action to remove a comment w/ macro response

Very quick and dirty implementation. No sanitisation of inputs. As yet
untested. You know the drill, reader: Use at your own peril.